Remove 'deep' restriction
When running the softAssert.include method, it checks if both the sample and the pattern are an array, if so it continues with a sample.include.
Howerver is one is not an instance of Array, then the method continues with a sample.deep.include - i don't remember why this is, but it kind of break the expected behaviour of the method.